unread count #742 attachment url #849

+++ server/api/broadcasts/broadcasts.utility.js
@@ -7,7 +7,7 @@ const TAG = 'api/broadcast/broadcasts.utility.js'
 const fs = require('fs')
 const path = require('path')
 
-function prepareSendAPIPayload (subscriberId, body, cb) {
+function prepareSendAPIPayload (subscriberId, body, isForLiveChat) {
   let payload = {}
   if (body.componentType === 'text' && !body.buttons) {
     payload = {
@@ -52,7 +52,7 @@ function prepareSendAPIPayload (subscriberId, body, cb) {
       }),
       'filedata': fileReaderStream
     }
-    deleteFile(body.fileurl)
+    if (!isForLiveChat) deleteFile(body.fileurl)
   } else if (body.componentType === 'card') {
     payload = {
       'recipient': JSON.stringify({
